StatisticDto types were changed

+++ app/src/main/java/com/betkey/network/models/StatisticDto.kt
@@ -2,22 +2,22 @@ package com.betkey.network.models
 
 data class StatisticDto(
     val baseData: BaseData,
-    val cancelledSportBetsAmount: Int,
+    val cancelledSportBetsAmount: Double,
     val cancelledSportBetsCount: Int,
     val depositsCount: Int,
-    val depositsTotalAmount: Int,
+    val depositsTotalAmount: Double,
     val errors: List<Any>,
     val message: String,
-    val payoutSportBetsAmount: Int,
+    val payoutSportBetsAmount: Double,
     val payoutSportBetsCount: Int,
-    val revertedSportBetsAmount: Int,
+    val revertedSportBetsAmount: Double,
     val revertedSportBetsCount: Int,
-    val soldJackpotBetsAmount: Int,
+    val soldJackpotBetsAmount: Double,
     val soldJackpotBetsCount: Int,
-    val soldSportBetsAmount: Int,
+    val soldSportBetsAmount: Double,
     val soldSportBetsCount: Int,
     val status: Boolean,
     val wallet: Wallet,
     val withdrawalsCount: Int,
-    val withdrawalsTotalAmount: Int
+    val withdrawalsTotalAmount: Double
 )
